Move insert modifiers on new scenarios to plugin and reshuffle things
