+++ rmf_site_editor/src/site/modifier.rs
@@ -1,0 +1,272 @@
+/*
+ * Copyright (C) 2025 Open Source Robotics Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+*/
+
+use crate::{
+    site::{
+        Affiliation, ChangeCurrentScenario, CurrentScenario, IssueKey, NameInSite, Property,
+        ScenarioMarker, ScenarioModifiers, UpdateProperty,
+    },
+    Issue, ValidateWorkspace,
+};
+use bevy::{
+    ecs::{component::Mutable, hierarchy::ChildOf, system::SystemParam},
+    prelude::*,
+};
+use std::fmt::Debug;
+use uuid::Uuid;
+
+#[derive(Clone, Debug, Event)]
+pub struct AddModifier {
+    for_element: Entity,
+    modifier: Entity,
+    in_scenario: Entity,
+    to_root: bool,
+}
+
+impl AddModifier {
+    pub fn new(for_element: Entity, modifier: Entity, in_scenario: Entity) -> Self {
+        Self {
+            for_element,
+            modifier,
+            in_scenario,
+            to_root: false,
+        }
+    }
+
+    pub fn new_to_root(for_element: Entity, modifier: Entity, in_scenario: Entity) -> Self {
+        Self {
+            for_element,
+            modifier,
+            in_scenario,
+            to_root: true,
+        }
+    }
+}
+
+#[derive(Clone, Debug, Event)]
+pub struct RemoveModifier {
+    for_element: Entity,
+    in_scenario: Entity,
+}
+
+impl RemoveModifier {
+    pub fn new(for_element: Entity, in_scenario: Entity) -> Self {
+        Self {
+            for_element,
+            in_scenario,
+        }
+    }
+}
+
+#[derive(SystemParam)]
+pub struct GetModifier<'w, 's, T: Component<Mutability = Mutable> + Clone + Default> {
+    pub scenarios: Query<
+        'w,
+        's,
+        (
+            &'static ScenarioModifiers<Entity>,
+            &'static Affiliation<Entity>,
+        ),
+        With<ScenarioMarker>,
+    >,
+    pub modifiers: Query<'w, 's, &'static T>,
+}
+
+impl<'w, 's, T: Component<Mutability = Mutable> + Clone + Default> GetModifier<'w, 's, T> {
+    pub fn get(&self, scenario: Entity, element: Entity) -> Option<&T> {
+        let mut modifier: Option<&T> = None;
+        let mut scenario_entity = scenario;
+        while modifier.is_none() {
+            let Ok((scenario_modifiers, scenario_parent)) = self.scenarios.get(scenario_entity)
+            else {
+                break;
+            };
+            if let Some(target_modifier) = scenario_modifiers
+                .get(&element)
+                .and_then(|e| self.modifiers.get(*e).ok())
+            {
+                modifier = Some(target_modifier);
+                break;
+            }
+
+            if let Some(parent_entity) = scenario_parent.0 {
+                scenario_entity = parent_entity;
+            } else {
+                // Modifier does not exist in the current scenario tree
+                break;
+            }
+        }
+        modifier
+    }
+}
+
+pub trait Modifier<T: Property>: Component<Mutability = Mutable> + Debug + Default + Clone {
+    /// This system retrieves the property values for this element's modifier, if any
+    fn get(&self) -> Option<T>;
+
+    /// This system climbs up the scenario tree to retrieve the inherited property
+    /// value for this element, if any.
+    fn retrieve_inherited(
+        &self,
+        for_element: Entity,
+        in_scenario: Entity,
+        get_modifier: &GetModifier<Self>,
+    ) -> Option<T> {
+        let mut parent_value: Option<T> = None;
+        let mut target_scenario = in_scenario;
+        while parent_value.is_none() {
+            let Some(parent_entity) = get_modifier
+                .scenarios
+                .get(target_scenario)
+                .ok()
+                .and_then(|(_, p)| p.0)
+            else {
+                break;
+            };
+
+            if let Some(modifier) = get_modifier.get(parent_entity, for_element) {
+                parent_value = modifier.get();
+            }
+            target_scenario = parent_entity;
+        }
+        parent_value
+    }
+
+    /// Inserts a new modifier for an element in the specified scenario. This is triggered
+    /// when property T is newly added to an element.
+    fn insert(_for_element: Entity, _in_scenario: Entity, _value: T, _world: &mut World) {}
+
+    /// Inserts new modifiers elements in a newly added root scenario.
+    fn insert_on_new_scenario(_in_scenario: Entity, _world: &mut World) {}
+}
+
+/// Handles additions and removals of scenario modifiers
+pub fn handle_scenario_modifiers(
+    mut commands: Commands,
+    mut change_current_scenario: EventWriter<ChangeCurrentScenario>,
+    mut add_modifier: EventReader<AddModifier>,
+    mut remove_modifier: EventReader<RemoveModifier>,
+    mut scenarios: Query<
+        (&mut ScenarioModifiers<Entity>, &Affiliation<Entity>),
+        With<ScenarioMarker>,
+    >,
+    current_scenario: Res<CurrentScenario>,
+) {
+    for remove in remove_modifier.read() {
+        let Ok((mut scenario_modifiers, _)) = scenarios.get_mut(remove.in_scenario) else {
+            continue;
+        };
+        if let Some(modifier) = scenario_modifiers.remove(&remove.for_element) {
+            commands.entity(modifier).despawn();
+        }
+
+        if current_scenario.0.is_some_and(|e| e == remove.in_scenario) {
+            change_current_scenario.write(ChangeCurrentScenario(remove.in_scenario));
+        };
+    }
+
+    for add in add_modifier.read() {
+        let scenario_entity = if add.to_root {
+            let mut target_scenario = add.in_scenario;
+            let mut root_scenario: Option<Entity> = None;
+            while root_scenario.is_none() {
+                let Ok((_, parent_scenario)) = scenarios.get(target_scenario) else {
+                    break;
+                };
+                if let Some(parent_entity) = parent_scenario.0 {
+                    target_scenario = parent_entity;
+                } else {
+                    root_scenario = Some(target_scenario);
+                    break;
+                }
+            }
+            if let Some(root_scenario_entity) = root_scenario {
+                root_scenario_entity
+            } else {
+                error!("No root scenario found for the current scenario tree!");
+                continue;
+            }
+        } else {
+            add.in_scenario
+        };
+
+        let Ok((mut scenario_modifiers, _)) = scenarios.get_mut(scenario_entity) else {
+            continue;
+        };
+        // If a modifier entity already exists, we assume it's meant to replace the current modifier.
+        // Despawn incoming modifier entity. Note that this erases any Recall data
+        if let Some(current_modifier) = scenario_modifiers.get(&add.for_element) {
+            commands.entity(*current_modifier).despawn();
+        }
+
+        commands
+            .entity(add.modifier)
+            .insert(Affiliation(Some(add.for_element)))
+            .insert(ChildOf(scenario_entity));
+        scenario_modifiers.insert(add.for_element, add.modifier);
+
+        commands.trigger(UpdateProperty::new(add.for_element, add.in_scenario));
+    }
+}
+
+/// Unique UUID to identify issue of missing root scenario modifiers
+pub const MISSING_ROOT_MODIFIER_ISSUE_UUID: Uuid =
+    Uuid::from_u128(0x98df792d3de44d26b126a9335f9e743au128);
+
+pub fn check_for_missing_root_modifiers<M: Component<Mutability = Mutable>>(
+    mut commands: Commands,
+    mut validate_events: EventReader<ValidateWorkspace>,
+    scenarios: Query<
+        (
+            &ScenarioModifiers<Entity>,
+            &NameInSite,
+            &Affiliation<Entity>,
+        ),
+        With<ScenarioMarker>,
+    >,
+    elements: Query<(Entity, Option<&NameInSite>), With<M>>,
+) {
+    for root in validate_events.read() {
+        for (scenario_modifiers, scenario_name, parent_scenario) in scenarios.iter() {
+            if parent_scenario.0.is_some() {
+                continue;
+            }
+            for (element, element_name) in elements.iter() {
+                if !scenario_modifiers.contains_key(&element) {
+                    let name = element_name
+                        .map(|name| name.0.clone())
+                        .unwrap_or(element.index().to_string());
+                    let issue = Issue {
+                        key: IssueKey {
+                            entities: [element].into(),
+                            kind: MISSING_ROOT_MODIFIER_ISSUE_UUID,
+                        },
+                        brief: format!(
+                            "Modifier for element {:?} is missing in root scenario {:?}",
+                            name, scenario_name.0
+                        ),
+                        hint: "Toggle the scenario properties for this element in specified scenario tree. \
+                               The editor will append a modifier with fallback values for this element."
+                            .to_string(),
+                    };
+                    let issue_id = commands.spawn(issue).id();
+                    commands.entity(**root).add_child(issue_id);
+                }
+            }
+        }
+    }
+}
